# 01/spec/api.yml
openapi: 3.0.3
info:
  title: 力麗飯店 LINE OA CRM API
  description: 多渠道會員管理與行銷平台 API，整合 LINE、Facebook Messenger 和 Webchat。
  version: 0.6.0
servers:
  - url: /api/v1
    description: API v1

# ============================================================================
# 既有功能 API（根據現有代碼與 Feature 規格）
# ============================================================================
paths:
...
  /line/message-notify:
    post:
      summary: LINE App 通知後端有新訊息（內部 webhook）
      operationId: notifyNewMessage
      tags: [LINE 訊息通知]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageNotification'
      responses:
        '200':
          description: 通知成功

  # ============================================================================
  # 新功能 API（FAQ 模組 + AI 聊天機器人）
  # Feature: faq_category_management, faq_rule_management, ai_chatbot_integration
  # ============================================================================

  # --------------------------------------------------------------------------
  # FAQ 大分類管理
  # Feature: faq_category_management
  # --------------------------------------------------------------------------
...
components:
  schemas:
...
    MessageNotification:
      type: object
      properties:
        member_id:
          type: integer
        message_type:
          type: string
        content:
          type: string
        channel:
          type: string

    # --- FAQ 管理（新功能）---
...